backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mongoengine==0.27.0
pymongo==4.7.2
django-cors-headers==4.3.1
python-dotenv==1.0.1
requests==2.32.3
celery==5.4.0
redis==5.0.4
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
